smart-dining-backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiofiles>=23.2.1
pillow>=10.2.0
numpy>=1.26.0
httpx>=0.25.0
# YOLOv13 dependencies
torch>=2.1.0
torchvision>=0.16.0
ultralytics>=8.1.0
